Neural Network
Neural Network for a basic binary classifier, choosing between a cat or not cat looking at the photos. 
The code will run with you pass in the correct data and download all modules but there are a few things to do to set up correct enviroment: email uenishia@bu.edu if you want to actually run the code. 
However, this is more for a visualization / step by step on how to create a very basic linear regression neural network with no hidden layers.